backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
PyPDF2==3.0.1
groq==0.4.2
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
fastapi-cache2==0.2.1
fastapi-limiter==0.1.6
redis==5.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-magic==0.4.27
aiofiles==23.2.1
aiohttp==3.9.3
tenacity==8.2.3
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
openai==1.12.0 